requirements.txt
joblib==1.1.1
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
streamlit==1.21.0
